Unit test for #196; no fix yet
